refactor: rename dashboard infrastructure directory
