+++ src/Service.Dashboard/src/Infrastructure/HostedServices/FileSystemWatcherBackgroundService.cs
