+++ src/Service.Dashboard/src/Infrastructure/Services/FileSystemWatcherService.cs
