+++ src/Service.Dashboard/src/Infrastructure/Setup.cs
@@ -10,13 +10,8 @@ using Giantnodes.Service.Dashboard.Domain.Aggregates.Entries.Files.Repositories;
 using Giantnodes.Service.Dashboard.Domain.Aggregates.Entries.Repositories;
 using Giantnodes.Service.Dashboard.Domain.Aggregates.Libraries.Repositories;
 using Giantnodes.Service.Dashboard.Domain.Services;
-using Giantnodes.Service.Dashboard.Infrastructure.Aggregates.Directories;
-using Giantnodes.Service.Dashboard.Infrastructure.Aggregates.EncodeProfiles;
-using Giantnodes.Service.Dashboard.Infrastructure.Aggregates.Encodes;
-using Giantnodes.Service.Dashboard.Infrastructure.Aggregates.Entries;
-using Giantnodes.Service.Dashboard.Infrastructure.Aggregates.Files;
-using Giantnodes.Service.Dashboard.Infrastructure.Aggregates.Libraries;
 using Giantnodes.Service.Dashboard.Infrastructure.HostedServices;
+using Giantnodes.Service.Dashboard.Infrastructure.Repositories;
 using Giantnodes.Service.Dashboard.Infrastructure.Services;
 using Giantnodes.Service.Dashboard.Persistence.DbContexts;
 using Microsoft.Extensions.DependencyInjection;
