+++ src/Service.Dashboard/tests/Giantnodes.Service.Dashboard.Application.Components.Tests/Libraries/Commands/LibraryCreateConsumerTests.cs
@@ -2,9 +2,6 @@
 using System.IO.Abstractions.TestingHelpers;
 using Giantnodes.Service.Dashboard.Application.Components.Libraries.Commands;
 using Giantnodes.Service.Dashboard.Application.Contracts.Libraries.Commands;
-using Giantnodes.Service.Dashboard.Domain;
-using Giantnodes.Service.Dashboard.Domain.Aggregates.Libraries.Services;
-using Giantnodes.Service.Dashboard.Domain.Aggregates.Libraries.Services.Impl;
 using Giantnodes.Service.Dashboard.Domain.Services;
 using Giantnodes.Service.Dashboard.Infrastructure.Services;
 using Giantnodes.Service.Dashboard.Persistence.DbContexts;
